stratum: Revert back definition of SubmitSharesSuccess
- change type of 'new_shares_sum' to 'u32' which is not consistent with
  the Stratum V2 specification but it does not break backward
  compatibility with old BOSminer
- the issue must be fixed in a different way

+++ stratum-proxy/src/translation.rs
@@ -819,7 +819,7 @@ impl V2ToV1Translation {
             channel_id: Self::CHANNEL_ID,
             last_seq_num: self.get_v2_submit_shares_seq_num(id)?,
             new_submits_accepted_count: 1,
-            new_shares_sum: new_shares,
+            new_shares_sum: new_shares as u32,
         };
 
         self.submit_share_response(success_msg, "SubmitSharesSuccess")
